+++ test/ck_tile/batched_transpose/batched_transpose_fp8.cpp
@@ -4,7 +4,7 @@
 
 int main()
 {
-    std::vector<std::vector<std::string>> test_cases = generate_test_cases("bf16");
+    std::vector<std::vector<std::string>> test_cases = generate_test_cases("fp8");
 
     return !run_test_cases<ck_tile::fp8_t>(test_cases);
 }
